requirements.txt
flask==3.1.0
beem==0.24.28
python-dotenv==1.1.0
gunicorn==23.0.0
